[EI-181] [web] refactor: Update import paths for CustomForecastIndicatorList
