+++ web/__tests__/components/numerical-guidance/custom-forecast-indicator/custom-forecast-indicator-list.test.tsx
@@ -3,7 +3,7 @@ import { SWRProviderWithoutCache } from '@/app/store/querys/swr-provider';
 import { resetMockDB } from '@/app/mocks/db';
 import { resetAllStore } from '@/app/store/stores/reset-store';
 import { useWorkspaceStore } from '@/app/store/stores/numerical-guidance/workspace.store';
-import CustomForecastIndicatorList from '@/app/ui/components/numerical-guidance/custom-forecast-indicator/custom-forecast-indicator-list';
+import CustomForecastIndicatorList from '@/app/ui/components/numerical-guidance/custom-forecast-indicator/custom-forecast-indicator-list/custom-forecast-indicator-list';
 import userEvent from '@testing-library/user-event';
 
 describe('CustomForecastIndicatorList', () => {
