+++ web/app/ui/components/numerical-guidance/custom-forecast-indicator/custom-forecast-indicator-list/custom-forecast-indicator-list-item.tsx
@@ -1,10 +1,10 @@
 import { CustomForecastIndicator } from '@/app/business/services/view-model/custom-forecast-indicator-view-model.service';
-import ListItem from '../../view/atom/list-item';
-import IconButton from '../../view/atom/icons/icon-button';
-import SelectableItem from '../../view/atom/selectable-item';
+import ListItem from '../../../view/atom/list-item';
+import IconButton from '../../../view/atom/icons/icon-button';
+import SelectableItem from '../../../view/atom/selectable-item';
 import { DotsHorizontalIcon } from '@heroicons/react/solid';
 import { useSelectedIndicatorBoardMetadata } from '@/app/business/hooks/indicator-board-metedata/use-selected-indicator-board-metadata-view-model.hook';
-import { useDialog } from '../../view/hooks/use-dialog.hook';
+import { useDialog } from '../../../view/hooks/use-dialog.hook';
 import { DIALOG_KEY } from '@/app/utils/keys/dialog-key';
 import { useSelectedCustomForecastIndicatorViewModel } from '@/app/business/hooks/custom-forecast-indicator/use-selected-custom-forecast-indicator-view-model';
