+++ web/app/ui/components/numerical-guidance/custom-forecast-indicator/custom-forecast-indicator-list/custom-forecast-indicator-list.tsx
@@ -2,10 +2,10 @@ import { useCustomForecastIndicatorListViewModel } from '@/app/business/hooks/cu
 import { useSelectedIndicatorBoardMetadata } from '@/app/business/hooks/indicator-board-metedata/use-selected-indicator-board-metadata-view-model.hook';
 import clsx from 'clsx';
 import React from 'react';
-import List from '../../view/molocule/list';
+import List from '../../../view/molocule/list';
 import { CustomForecastIndicator } from '@/app/business/services/view-model/custom-forecast-indicator-view-model.service';
 import CustomForecastIndicatorListItem from './custom-forecast-indicator-list-item';
-import Pending from '../../view/molocule/pending';
+import Pending from '../../../view/molocule/pending';
 
 const CustomForecastIndicatorList = React.memo(function CustomForecastIndicatorList() {
   const { customForecastIndicatorList, isPending } = useCustomForecastIndicatorListViewModel();
